more merge fixes
